server/server/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.7
Django==4.2.7
django-cors-headers==4.3.1
django-json-widget==1.1.1
django-prettyjson==0.4.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
djangorestframework-simplejwt==5.3.1
future==0.18.3
inflection==0.5.1
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
standardjson==0.3.1
tzdata==2023.3